[CC#31] (https://github.com/Components-care/backlog/issues/31) Changes for restructuring signature pad.
