+++ src/non-standalone/Dialog/SignPadDialog.tsx
