+++ src/non-standalone/index.ts
@@ -1,2 +1,3 @@
 export * from "./Dialog";
 export * from "./Portal";
+export { default as SignaturePad } from "./SignaturePad/SignaturePad";
