+++ src/standalone/index.ts
@@ -13,5 +13,4 @@ export * from "./LocalizedDateTimePickers";
 export { default as InfoBox } from "./InfoBox";
 export * from "./Icons";
 export * from "./UIKit";
-export { default as SignaturePad } from "./SignaturePad";
 export { default as Notifications } from "./Notifications";
